fix(stella-cli,stella-runtime): fleet_cmd splits its report half; main is green
`main` at `a7a46d3b6` fails `check-file-size`: `fleet_cmd.rs` is 1538 lines and
is not grandfathered, so the guard's remedy is a split and explicitly not a
baseline entry (#4821).

The four functions that turn a finished run into terminal output move to
`fleet_cmd::render` — `render_watch_line`, `truncate`, `print_dash_summary`,
`render_report`, and the `SUMMARY_CHARS` width they share. They are the group
in the file with no callers beyond the two print points and no dependency on
the run machinery: they read a finished report and write to stdout. The parent
drops to 1375 lines.

`SUMMARY_CHARS` is imported by the test module from `render` rather than
through the parent, because the parent does not use it and re-exporting it
only to satisfy a test is how an unused import becomes a warning nobody reads.

Two more reds on the same tree, composed here rather than raced on separate
branches — the repairs deadlock each other otherwise:

- **prose.** Four `deliberately` filler-adverbs over the per-pattern
  allowance, in `goal_wrapped/tests.rs`, `fleet_cmd/wrapped.rs` (two) and
  `stella-runtime/README.md`. Each sentence already gives its reason; the
  adverb was standing in front of it.
- **clippy.** `explicit_auto_deref` on `RepublishingDriver::new`'s registry
  argument. `&Arc<T>` coerces to `&T` on its own.

Evidence: `make guards-fast` exit 0 (`check-file-size` OK, `check-prose` OK);
`make lint CARGO_SCOPE="-p stella-cli"` exit 0; `make doc-warnings` over
stella-cli + stella-runtime exit 0; `cargo test -p stella-cli --bin stella`
2288 passed, 0 failed; `cargo fmt --check` clean.

Closes #4821.
